+++ tests/unit/contracts/test_independence.py
@@ -240,12 +240,18 @@ def test_independence_contract(shortest_chains, expected_invalid_chains):
             },
         ],
         shortest_chains=shortest_chains,
-
+        all_modules=[
+            'mypackage',
+            'mypackage.blue', 'mypackage.blue.alpha', 'mypackage.blue.beta',
+            'mypackage.blue.beta.foo',
+            'mypackage.green',
+            'mypackage.yellow', 'mypackage.yellow.gamma', 'mypackage.yellow.delta',
+        ]
     )
     contract = IndependenceContract(
         name='Independence contract',
         session_options={
-            'root_package_name': 'mypackage',
+            'root_package': 'mypackage',
         },
         contract_options={
             'modules': (
@@ -315,11 +321,14 @@ def test_ignore_imports(ignore_imports, is_kept):
         shortest_chains={
             ('a', 'b'): ('a', 'indirect', 'b'),
         },
+        all_modules=[
+            'mypackage', 'mypackage.a', 'mypackage.b', 'mypackage.indirect',
+        ]
     )
     contract = IndependenceContract(
         name='Independence contract',
         session_options={
-            'root_package_name': 'mypackage',
+            'root_package': 'mypackage',
         },
         contract_options={
             'modules': (
@@ -342,7 +351,7 @@ def test_render_broken_contract():
     contract = IndependenceContract(
         name='Independence contract',
         session_options={
-            'root_package_name': 'mypackage',
+            'root_package': 'mypackage',
         },
         contract_options={
             'modules': [
@@ -423,3 +432,34 @@ def test_render_broken_contract():
 
         """
     )
+
+
+def test_missing_module():
+    graph = FakeGraph(
+        root_package_name='mypackage',
+        all_modules=[
+            'mypackage',
+            'mypackage.foo',
+        ]
+    )
+
+    contract = IndependenceContract(
+        name='Independence contract',
+        session_options={
+            'root_package': 'mypackage',
+        },
+        contract_options={
+            'modules': [
+                'mypackage.foo',
+                'mypackage.bar',
+            ],
+        },
+    )
+
+    with pytest.raises(
+        ValueError,
+        match=(
+            "Module 'mypackage.bar' does not exist."
+        )
+    ):
+        contract.check(graph=graph)